# Run `findall(rows)` only if `rows` are not all true