PR that resulted from #2724 disussion.
Addition of `_findall(rows::AbstractVector{Bool})`
to not run `findall(rows)` if `all(rows .== true)` in
`function SubDataFrame(parent::DataFrame, rows::AbstractVector{Bool}, cols)`

I'm not sure if `src/other/utils.jl` is a good location for this function. 

------------------

Benchmarks:
```julia
df = DataFrame(
    a = rand(1000000),
    b = allowmissing(rand(1000000)),
    c = [allowmissing(rand(1000000 - 1)); missing],
    d = repeat([1:9; missing], 100000)
)
```

Main branch
```
 julia> @btime dropmissing(df, :a, view=true);
  2.040 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :b, view=true);
  2.071 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :c, view=true);
  2.098 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :d, view=true);
  1.911 ms (15 allocations: 6.99 MiB)

julia> @btime dropmissing(df, view=true);
  7.677 ms (37 allocations: 7.00 MiB)
```

Main + this fix
```
julia> @btime dropmissing(df, :a, view=true);
  388.342 μs (13 allocations: 126.64 KiB)

julia> @btime dropmissing(df, :b, view=true);
  399.634 μs (13 allocations: 126.64 KiB)

julia> @btime dropmissing(df, :c, view=true);
  2.103 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :d, view=true);
  1.897 ms (15 allocations: 6.99 MiB)

julia> @btime dropmissing(df, view=true);
  7.679 ms (37 allocations: 7.00 MiB)
  ```
  
After `completecases` fix #2726 
```
julia> @btime dropmissing(df, :a, view=true);
  1.883 ms (10 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :b, view=true);
  2.075 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :c, view=true);
  2.095 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :d, view=true);
  1.905 ms (15 allocations: 6.99 MiB)

julia> @btime dropmissing(df, view=true);
  2.351 ms (28 allocations: 7.12 MiB)
```

After `completecases` fix #2726 + this fix

```
julia> @btime dropmissing(df, :a, view=true);
  251.012 μs (8 allocations: 122.38 KiB)

julia> @btime dropmissing(df, :b, view=true);
  399.373 μs (13 allocations: 126.64 KiB)

julia> @btime dropmissing(df, :c, view=true);
  2.109 ms (15 allocations: 7.75 MiB)

julia> @btime dropmissing(df, :d, view=true);
  1.903 ms (15 allocations: 6.99 MiB)

julia> @btime dropmissing(df, view=true);
  2.355 ms (28 allocations: 7.12 MiB)
```

------------------

Summary:

* #2726 Reduces time of `dropmissing(df, view=true)`
* This fix reduces average time for `dropmissing(df, :a, view=true)`. In the worst case scenario (as with `:c` column) where there is only one missing in the end it is a bit slower, but I think that it is very rarely case. Usually we'll find missings/falses earlier.