@pstorozenko 2021-05-09
I think this is more what you were talking about :)
I took `Base.findall(B::BitArray)` function and rewrote/adapted it for our needs.

In the first `while true` we process `B` in the optimized way.
After it we either returned solution or materialized result `I`.

In the second `while true` we process `B` in _almost_ the same way as in `Base.findall(B::BitArray)`.

I've tried to comment code a lot as operations are not obvious.
I've used one line ifs like `Bi == length(Bc) && return start:length(B)` because it makes code much cleaner, but I don't know if you like it.
Tell me what do you think about this whole approach.

In general this solution is sometimes a bit slower in case of random `B`.
This is because `Base.findall(B::BitArray)` does not try to _jump_ over blocks with trues only.
In case of bigger blocks of trues it is faster.
Of course in case of single, continuous block of trues it is much, much faster.

Bellow I attach some benchmarks:

```julia
BD = OrderedDict(
    "T Big" => trues(100000),
    "F Big" => falses(100000),
    "T64 F64" => [trues(64); falses(64)],
    "F64 T64" => [falses(64); trues(64)],
    "F80 T100" => [falses(85); trues(100)],
    "F256 T32" => [falses(256); trues(32)],
    "F260 T32" => [falses(260); trues(32)],
    "TF Big" => [trues(100000); falses(100000)],
    "FT Big" => [falses(100000);trues(100000)] ,

    # some edge cases
    "TFT small" => [trues(85); falses(100); trues(85)],
    "FTFFT small" => [falses(64 + 32); trues(32); falses(128); trues(32)],
    "TFTF small" => [falses(64); trues(64); falses(64); trues(64)],
    "TFT small" => [trues(64); falses(10); trues(100)],

    "FTF Big" => [falses(8500); trues(100000); falses(65000)],
    "TFT Big" => [trues(8500); falses(100000); trues(65000)],
    "FTFTFTF Big" => [falses(65000); trues(65000); falses(65000); trues(65000); falses(65000); trues(65000); falses(65000)],

    "FTFR small" => [falses(85); trues(100); falses(65); rand([true, false], 20)],
    "R Big" => BitVector(rand([true, false], 200000)),
    "RF Big" => [BitVector(rand([true, false], 100000)) ; falses(100000)],
    "RT Big" => [BitVector(rand([true, false], 100000)) ; trues(100000)],
    "FTFR Big" => [falses(65000);  trues(65000);  falses(65000); rand([true, false], 20000)],
    "T256 R100" => [trues(256);  rand([true, false], 100)],
    "F256 R100" => [falses(256); rand([true, false], 100)],
)

for (l, B) in BD
    println("Processing $l")
    @show Base.findall(B) == DataFrames._findall(B)
    @btime Base.findall($B)
    @btime DataFrames._findall($B)
end
```

```
Processing T Big
Base.findall(B) == DataFrames._findall(B) = true
  109.098 μs (2 allocations: 781.33 KiB)
  398.537 ns (0 allocations: 0 bytes)
Processing F Big
Base.findall(B) == DataFrames._findall(B) = true
  430.095 ns (1 allocation: 80 bytes)
  421.095 ns (1 allocation: 80 bytes)
Processing T64 F64
Base.findall(B) == DataFrames._findall(B) = true
  104.586 ns (1 allocation: 624 bytes)
  41.959 ns (1 allocation: 80 bytes)
Processing F64 T64
Base.findall(B) == DataFrames._findall(B) = true
  104.734 ns (1 allocation: 624 bytes)
  41.498 ns (1 allocation: 80 bytes)
Processing F80 T100
Base.findall(B) == DataFrames._findall(B) = true
  143.548 ns (1 allocation: 896 bytes)
  46.941 ns (1 allocation: 80 bytes)
Processing F256 T32
Base.findall(B) == DataFrames._findall(B) = true
  73.466 ns (1 allocation: 336 bytes)
  45.835 ns (1 allocation: 80 bytes)
Processing F260 T32
Base.findall(B) == DataFrames._findall(B) = true
  73.387 ns (1 allocation: 336 bytes)
  45.847 ns (1 allocation: 80 bytes)
Processing TF Big
Base.findall(B) == DataFrames._findall(B) = true
  105.403 μs (2 allocations: 781.33 KiB)
  1.467 μs (1 allocation: 80 bytes)
Processing FT Big
Base.findall(B) == DataFrames._findall(B) = true
  105.608 μs (2 allocations: 781.33 KiB)
  1.987 μs (1 allocation: 80 bytes)
Processing TFT small
Base.findall(B) == DataFrames._findall(B) = true
  217.728 ns (1 allocation: 1.45 KiB)
  188.086 ns (2 allocations: 1.41 KiB)
Processing FTFFT small
Base.findall(B) == DataFrames._findall(B) = true
  106.897 ns (1 allocation: 624 bytes)
  112.246 ns (2 allocations: 624 bytes)
Processing TFTF small
Base.findall(B) == DataFrames._findall(B) = true
  215.025 ns (1 allocation: 1.14 KiB)
  160.212 ns (2 allocations: 1.14 KiB)
Processing FTF Big
Base.findall(B) == DataFrames._findall(B) = true
  127.988 μs (2 allocations: 781.33 KiB)
  1.767 μs (1 allocation: 80 bytes)
Processing TFT Big
Base.findall(B) == DataFrames._findall(B) = true
  95.470 μs (2 allocations: 574.33 KiB)
  66.360 μs (2 allocations: 574.39 KiB)
Processing FTFTFTF Big
Base.findall(B) == DataFrames._findall(B) = true
  253.018 μs (2 allocations: 1.49 MiB)
  194.054 μs (2 allocations: 1.49 MiB)
Processing FTFR small
Base.findall(B) == DataFrames._findall(B) = true
  189.834 ns (1 allocation: 1008 bytes)
  153.568 ns (2 allocations: 976 bytes)
Processing R Big
Base.findall(B) == DataFrames._findall(B) = true
  153.825 μs (2 allocations: 778.89 KiB)
  119.013 μs (2 allocations: 778.95 KiB)
Processing RF Big
Base.findall(B) == DataFrames._findall(B) = true
  78.821 μs (2 allocations: 389.64 KiB)
  59.672 μs (2 allocations: 389.70 KiB)
Processing RT Big
Base.findall(B) == DataFrames._findall(B) = true
  204.643 μs (2 allocations: 1.15 MiB)
  165.110 μs (2 allocations: 1.15 MiB)
Processing FTFR Big
Base.findall(B) == DataFrames._findall(B) = true
  100.556 μs (2 allocations: 586.64 KiB)
  72.579 μs (2 allocations: 586.70 KiB)
Processing T256 R100
Base.findall(B) == DataFrames._findall(B) = true
  659.301 ns (1 allocation: 2.56 KiB)
  519.550 ns (2 allocations: 2.58 KiB)
Processing F256 R100
Base.findall(B) == DataFrames._findall(B) = true
  120.717 ns (1 allocation: 544 bytes)
  131.407 ns (2 allocations: 528 bytes)
```