@bkamins 2021-05-10
Yes - I have checked this also :). It seems that probably the compiler somehow optimizes these lines out (not sure though).