@pstorozenko 2021-05-14
@nalimilan What do you mean by _fix the uncovered line_?
Do you mean code coverage? I'm sure all lines (except of `@assert false "should not be reached"`) are covered by tests (I tested with good old `@show`), but for some reason codecov is showing differently.