@bkamins 2021-05-14
> but for some reason codecov is showing differently.

I have also checked the logic that the important lines are covered.